--- a/commitlint.config.cjs
+++ b/commitlint.config.cjs
@@ -1,1 +1,7 @@
-module.exports = { extends: ['@commitlint/config-conventional'] };
+module.exports = {
+  extends: ['@commitlint/config-conventional'],
+  rules: {
+    'body-max-line-length': [2, 'always', 200],
+    'footer-max-line-length': [2, 'always', 200],
+  },
+}