… .claude/settings.local.json
… .zed/settings.json
… CHANGELOG.md
… CLAUDE.md
… OPTIMIZATIONS.md
… README.md
… TUNNEL_TESTS.md
… WRAP_ARCHITECTURE.md
… docs/AUDIT_SUMMARY.md
… docs/NEXT_STEPS.md
… package.json
… scripts/performance-test.cjs
… src/App.test.tsx
… src/components/Pacman.tsx
… src/components/PerformanceMonitor.tsx
… src/game/__tests__/movement.test.ts
… src/game/__tests__/wrap.test.ts
… src/game/grid.ts
… src/game/react/useGame.ts
… src/game/state.ts
… src/hooks/usePerformance.ts
… vite.config.ts